chore(src): rename page -> pages
