+++ src/pages/Login/index.ts
@@ -1,6 +1,6 @@
 import { just, Stream } from 'most';
 import { VNode, DOMSource } from '@motorcycle/dom';
-import { Sources } from '../../component/types';
+import { Sources } from '../../components/types';
 
 import { authenticate, AuthenticationSinks } from '../../higher-order-components/authenticate';
 import isolate from '@cycle/isolate';
