+++ src/pages/main.ts
